requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
mysql-connector-python==8.0.33
sqlalchemy==2.0.15
fastapi-cache2==0.2.0
redis==4.5.5
pydantic==1.10.7